# docs/swagger.yaml
basePath: /
definitions:
...
  handler.LogEventRequest:
    properties:
      event_type:
        allOf:
        - $ref: '#/definitions/model.EventType'
        example: play
      project_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
    type: object
...
  model.AnalyticsEvent:
    properties:
      created_at:
        type: string
      event_type:
        $ref: '#/definitions/model.EventType'
      id:
        type: string
      project_id:
        type: string
      user_id:
        type: string
    type: object
  model.Asset:
    properties:
      created_at:
        type: string
      id:
        type: string
      mime_type:
        type: string
      original_filename:
        type: string
      project_id:
        type: string
      size:
        type: integer
      stored_filename:
        type: string
    type: object
  model.EventType:
    enum:
    - play
    - click
    - impression
    type: string
    x-enum-varnames:
    - EventTypePlay
    - EventTypeClick
    - EventTypeImpression
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the backend API for the Playable Ads Manager Platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Playable Ads Manager Backend API
  version: "1.0"
paths:
  /analytics:
    get:
      description: Retrieves a list of all analytics events across all projects. This
        is a public endpoint.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AnalyticsEvent'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: List all analytics events
      tags:
      - analytics
    post:
      consumes:
      - application/json
      description: Logs an analytics event for a project owned by the user.
      parameters:
      - description: Analytics Event Info
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/handler.LogEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.AnalyticsEvent'
        "400":
          description: Invalid request body or event type
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Log an analytics event
      tags:
      - analytics
  /assets/{id}:
    get:
      description: Serves the asset file if it belongs to the requesting user.
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Invalid asset ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a specific asset file
      tags:
      - assets
...
  /projects/{id}/assets:
    get:
      description: Retrieves a list of all assets for a specific project.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Asset'
            type: array
        "400":
          description: Invalid project ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List assets in a project
      tags:
      - assets
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image or video file to a specific project.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Asset file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Asset'
        "400":
          description: Invalid project ID or file error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "413":
          description: File size exceeds limit
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload an asset to a project
      tags:
      - assets
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"